+++ packages/core/minos-microservice-aggregate/tests/test_aggregate/test_models/test_aggregates/test_broker.py
@@ -2,10 +2,6 @@ import unittest
 from typing import (
     Optional,
 )
-from unittest.mock import (
-    AsyncMock,
-    call,
-)
 
 from minos.aggregate import (
     Action,
@@ -15,7 +11,7 @@ from minos.aggregate import (
     ModelRef,
 )
 from minos.networks import (
-    BrokerMessageStrategy,
+    BrokerMessageV1,
 )
 from tests.utils import (
     Car,
@@ -26,99 +22,88 @@ from tests.utils import (
 
 class TestAggregate(MinosTestCase):
     async def test_create(self):
-        mock = AsyncMock()
-        self.broker_publisher.send = mock
-
         car = await Car.create(doors=3, color="blue")
 
+        observed = self.broker_publisher.messages
+
+        self.assertEqual(1, len(observed))
+        self.assertIsInstance(observed[0], BrokerMessageV1)
+        self.assertEqual("CarCreated", observed[0].topic)
         self.assertEqual(
-            [
-                call(
-                    AggregateDiff(
-                        uuid=car.uuid,
-                        name=Car.classname,
-                        version=1,
-                        action=Action.CREATE,
-                        created_at=car.created_at,
-                        fields_diff=FieldDiffContainer(
-                            [
-                                FieldDiff("doors", int, 3),
-                                FieldDiff("color", str, "blue"),
-                                FieldDiff("owner", Optional[ModelRef[Owner]], None),
-                            ]
-                        ),
-                    ),
-                    "CarCreated",
-                    strategy=BrokerMessageStrategy.MULTICAST,
-                )
-            ],
-            mock.call_args_list,
+            AggregateDiff(
+                uuid=car.uuid,
+                name=Car.classname,
+                version=1,
+                action=Action.CREATE,
+                created_at=car.created_at,
+                fields_diff=FieldDiffContainer(
+                    [
+                        FieldDiff("doors", int, 3),
+                        FieldDiff("color", str, "blue"),
+                        FieldDiff("owner", Optional[ModelRef[Owner]], None),
+                    ]
+                ),
+            ),
+            observed[0].content,
         )
 
     async def test_update(self):
-        mock = AsyncMock()
-        self.broker_publisher.send = mock
-
         car = await Car.create(doors=3, color="blue")
-        mock.reset_mock()
+        self.broker_publisher.messages.clear()
 
         await car.update(color="red")
 
+        observed = self.broker_publisher.messages
+
+        self.assertEqual(2, len(observed))
+        self.assertIsInstance(observed[0], BrokerMessageV1)
+        self.assertEqual("CarUpdated", observed[0].topic)
         self.assertEqual(
-            [
-                call(
-                    AggregateDiff(
-                        uuid=car.uuid,
-                        name=Car.classname,
-                        version=2,
-                        action=Action.UPDATE,
-                        created_at=car.updated_at,
-                        fields_diff=FieldDiffContainer([FieldDiff("color", str, "red")]),
-                    ),
-                    "CarUpdated",
-                    strategy=BrokerMessageStrategy.MULTICAST,
-                ),
-                call(
-                    AggregateDiff(
-                        uuid=car.uuid,
-                        name=Car.classname,
-                        version=2,
-                        action=Action.UPDATE,
-                        created_at=car.updated_at,
-                        fields_diff=FieldDiffContainer([FieldDiff("color", str, "red")]),
-                    ),
-                    "CarUpdated.color",
-                    strategy=BrokerMessageStrategy.MULTICAST,
-                ),
-            ],
-            mock.call_args_list,
+            AggregateDiff(
+                uuid=car.uuid,
+                name=Car.classname,
+                version=2,
+                action=Action.UPDATE,
+                created_at=car.updated_at,
+                fields_diff=FieldDiffContainer([FieldDiff("color", str, "red")]),
+            ),
+            observed[0].content,
+        )
+        self.assertIsInstance(observed[1], BrokerMessageV1)
+        self.assertEqual("CarUpdated.color", observed[1].topic)
+        self.assertEqual(
+            AggregateDiff(
+                uuid=car.uuid,
+                name=Car.classname,
+                version=2,
+                action=Action.UPDATE,
+                created_at=car.updated_at,
+                fields_diff=FieldDiffContainer([FieldDiff("color", str, "red")]),
+            ),
+            observed[1].content,
         )
 
     async def test_delete(self):
-        mock = AsyncMock()
-        self.broker_publisher.send = mock
-
         car = await Car.create(doors=3, color="blue")
-        mock.reset_mock()
+        self.broker_publisher.messages.clear()
 
         await car.delete()
 
+        observed = self.broker_publisher.messages
+
+        self.assertEqual(1, len(observed))
+        self.assertIsInstance(observed[0], BrokerMessageV1)
+        self.assertEqual("CarDeleted", observed[0].topic)
         self.assertEqual(
-            [
-                call(
-                    AggregateDiff(
-                        uuid=car.uuid,
-                        name=Car.classname,
-                        version=2,
-                        action=Action.DELETE,
-                        created_at=car.updated_at,
-                        fields_diff=FieldDiffContainer.empty(),
-                    ),
-                    "CarDeleted",
-                    strategy=BrokerMessageStrategy.MULTICAST,
-                )
-            ],
-            mock.call_args_list,
+            AggregateDiff(
+                uuid=car.uuid,
+                name=Car.classname,
+                version=2,
+                action=Action.DELETE,
+                created_at=car.updated_at,
+                fields_diff=FieldDiffContainer.empty(),
+            ),
+            observed[0].content,
         )