+++ packages/core/minos-microservice-cqrs/docs/conf.py
@@ -19,9 +19,14 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('..'))
 
-from minos import cqrs
+sys.path.insert(0, os.path.abspath(".."))
+
+import sphinx_rtd_theme
+
+from minos import (
+    cqrs,
+)
 
 # -- General configuration ---------------------------------------------
 
@@ -32,31 +37,30 @@ from minos import cqrs
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
 
-import sphinx_rtd_theme
 
 extensions = [
     "sphinxcontrib.apidoc",
-    'sphinx.ext.autodoc',
+    "sphinx.ext.autodoc",
     "sphinx_autodoc_typehints",
     "sphinx.ext.viewcode",
     "sphinx_rtd_theme",
     "m2r2",
 ]
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
-source_suffix = ['.rst', '.md']
+source_suffix = [".rst", ".md"]
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # General information about the project.
-project = 'Minos Microservice CQRS'
+project = "Minos Microservice CQRS"
 copyright = "2021, Clariteia"
-author = "Clariteia Devs"
+author = "Minos Framework Devs"
 
 # The version info for the project you're documenting, acts as replacement
 # for |version| and |release|, also used in various other places throughout
@@ -77,10 +81,10 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
+pygments_style = "sphinx"
 
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = False
@@ -92,7 +96,7 @@ todo_include_todos = False
 # a list of builtin themes.
 #
 
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
 
 # Theme options are theme-specific and customize the look and feel of a
 # theme further.  For a list of options available for each theme, see the
@@ -119,7 +123,7 @@ html_static_path = ["_static"]
 # -- Options for HTMLHelp output ---------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'minosdoc'
+htmlhelp_basename = "minosdoc"
 
 
 # -- Options for LaTeX output ------------------------------------------
@@ -128,15 +132,12 @@ latex_elements = {
     # The paper size ('letterpaper' or 'a4paper').
     #
     # 'papersize': 'letterpaper',
-
     # The font size ('10pt', '11pt' or '12pt').
     #
     # 'pointsize': '10pt',
-
     # Additional stuff for the LaTeX preamble.
     #
     # 'preamble': '',
-
     # Latex figure (float) alignment
     #
     # 'figure_align': 'htbp',
@@ -146,9 +147,7 @@ latex_elements = {
 # (source start file, target name, title, author, documentclass
 # [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'minos.tex',
-     'Minos Microservice CQRS Documentation',
-     'Clariteia Devs', 'manual'),
+    (master_doc, "minos.tex", "Minos Microservice CQRS Documentation", "Minos Framework Devs", "manual"),
 ]
 
 
@@ -156,11 +155,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, 'minos',
-     'Minos Microservice CQRS Documentation',
-     [author], 1)
-]
+man_pages = [(master_doc, "minos", "Minos Microservice CQRS Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output ----------------------------------------
@@ -169,12 +164,15 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'minos',
-     'Minos Microservice CQRS Documentation',
-     author,
-     'minos',
-     'One line description of project.',
-     'Miscellaneous'),
+    (
+        master_doc,
+        "minos",
+        "Minos Microservice CQRS Documentation",
+        author,
+        "minos",
+        "One line description of project.",
+        "Miscellaneous",
+    ),
 ]
 
 # "apidoc" extension