+++ packages/core/minos-microservice-networks/minos/networks/discovery/clients/abc.py
@@ -25,7 +25,7 @@ class DiscoveryClient(ABC):
 
     @property
     def route(self) -> str:
-        """Get the full http route to the repository.
+        """Get the full http route to the discovery.
 
         :return: An ``str`` value.
         """
@@ -42,7 +42,7 @@ class DiscoveryClient(ABC):
         retry_tries: int = 3,
         retry_delay: float = 5,
     ) -> None:
-        """ Perform a subscription query.
+        """Perform a subscription query.
 
         :param host: The ip of the microservice to be subscribed.
         :param port: The port of the microservice to be subscribed.
@@ -82,7 +82,7 @@ class DiscoveryClient(ABC):
 
     @abstractmethod
     async def unsubscribe(self, name: str, retry_tries: int = 3, retry_delay: float = 5) -> None:
-        """ Perform an unsubscribe query.
+        """Perform an unsubscribe query.
 
         :param name: The name of the microservice to be unsubscribed.
         :param retry_tries: Number of attempts before raising a failure exception.