+++ packages/core/minos-microservice-networks/tests/test_networks/test_rest/utils.py
@@ -31,27 +31,29 @@ from minos.common import (
 
 
 def json_mocked_request(data: Any, **kwargs) -> web.Request:
-    """For testng purposes. """
+    """For testng purposes."""
     return mocked_request(json.dumps(data).encode(), content_type="application/json", **kwargs)
 
 
 def form_mocked_request(data: dict[str, Any], **kwargs) -> web.Request:
-    """For testng purposes. """
-    return mocked_request(urlencode(data, doseq=True).encode(), content_type="application/x-www-form-encoded", **kwargs)
+    """For testng purposes."""
+    return mocked_request(
+        urlencode(data, doseq=True).encode(), content_type="application/x-www-form-urlencoded", **kwargs
+    )
 
 
 def avro_mocked_request(data: Any, schema: Any, **kwargs) -> web.Request:
-    """For testng purposes. """
+    """For testng purposes."""
     return mocked_request(MinosAvroProtocol.encode(data, schema), content_type="avro/binary", **kwargs)
 
 
 def text_mocked_request(data: str, **kwargs) -> web.Request:
-    """For testng purposes. """
+    """For testng purposes."""
     return mocked_request(data.encode(), content_type="text/plain", **kwargs)
 
 
 def bytes_mocked_request(data: bytes, **kwargs) -> web.Request:
-    """For testng purposes. """
+    """For testng purposes."""
     return mocked_request(data, content_type="application/octet-stream", **kwargs)
 
 
@@ -72,7 +74,7 @@ def mocked_request(
         headers = headers.copy()
 
     if user is not None:
-        headers["User"] = str(user)
+        headers["user"] = str(user)
 
     if content_type is not None:
         headers["Content-Type"] = content_type