+++ packages/core/minos-microservice-saga/tests/test_saga/test_messages.py
@@ -4,7 +4,8 @@ from uuid import (
 )
 
 from minos.networks import (
-    BrokerMessage,
+    BrokerMessageV1,
+    BrokerMessageV1Payload,
 )
 from minos.saga import (
     SagaRequest,
@@ -48,13 +49,17 @@ class TestSagaResponse(unittest.IsolatedAsyncioTestCase):
     def test_from_message(self):
         expected = SagaResponse(56, {"ticket", "product"}, uuid=self.uuid)
         observed = SagaResponse.from_message(
-            BrokerMessage("", 56, headers={"saga": str(self.uuid), "related_services": "ticket,product"})
+            BrokerMessageV1(
+                "", BrokerMessageV1Payload(56, headers={"saga": str(self.uuid), "related_services": "ticket,product"})
+            )
         )
         self.assertEqual(expected, observed)
 
     def test_from_message_empty_related_services(self):
         expected = SagaResponse(56, uuid=self.uuid)
-        observed = SagaResponse.from_message(BrokerMessage("", 56, headers={"saga": str(self.uuid)}))
+        observed = SagaResponse.from_message(
+            BrokerMessageV1("", BrokerMessageV1Payload(56, headers={"saga": str(self.uuid)}))
+        )
         self.assertEqual(expected, observed)
 
     async def test_content(self):